post fixed changes
git-svn-id: https://fheroes2.svn.sourceforge.net/svnroot/fheroes2/trunk@742 4f904287-c015-0410-838d-95d3a2656228

+++ fheroes2/src/fheroes2/dialog/dialog.h
@@ -133,7 +133,7 @@ private:
 class FrameBorder
 {
 public:
-    FrameBorder(bool fade=true, int width = 640, int height = 480);
+    FrameBorder(bool fade=true, u16 width = 640, u16 height = 480);
     ~FrameBorder();
 
     const Rect & GetArea(void) const{ return area; };
